refactor: codemirror split
