+++ packages/editor/src/extensions/CodeMIrror/codemirror-utils.ts
@@ -12,6 +12,7 @@ import { Selection } from '@remirror/pm/state'
  */
 export function arrowHandler(dir: 'left' | 'right' | 'up' | 'down'): CommandFunction {
   return ({ dispatch, view, tr }) => {
+    console.log('arrowHandler', dir, tr.selection.empty, view?.endOfTextblock(dir))
     if (!view) {
       return false
     }
